phase-3/mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP Server for Todo Task Management"
authors = [{name = "Developer", email = "dev@example.com"}]
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0"
]
requires-python = ">=3.13"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"